+++ code-pushup.preset.ts
@@ -8,8 +8,9 @@ import coveragePlugin, {
 } from './packages/plugin-coverage/src/index.js';
 import eslintPlugin, {
   eslintConfigFromAllNxProjects,
-  eslintConfigFromNxProject,
 } from './packages/plugin-eslint/src/index.js';
+import type { ESLintTarget } from './packages/plugin-eslint/src/lib/config.js';
+import { nxProjectsToConfig } from './packages/plugin-eslint/src/lib/nx/projects-to-config.js';
 import jsPackagesPlugin from './packages/plugin-js-packages/src/index.js';
 import jsDocsPlugin from './packages/plugin-jsdocs/src/index.js';
 import type { JsDocsPluginTransformedConfig } from './packages/plugin-jsdocs/src/lib/config.js';
@@ -156,15 +157,27 @@ export const jsDocsCoreConfig = (
   ),
 });
 
+export async function eslintConfigFromPublishableNxProjects(): Promise<
+  ESLintTarget[]
+> {
+  const { createProjectGraphAsync } = await import('@nx/devkit');
+  const projectGraph = await createProjectGraphAsync({ exitOnError: false });
+  return nxProjectsToConfig(
+    projectGraph,
+    project => project.tags?.includes('publishable') ?? false,
+  );
+}
+
 export const eslintCoreConfigNx = async (
   projectName?: string,
 ): Promise<CoreConfig> => ({
   plugins: [
-    await eslintPlugin(
-      await (projectName
-        ? eslintConfigFromNxProject(projectName)
-        : eslintConfigFromAllNxProjects()),
-    ),
+    projectName
+      ? await eslintPlugin({
+          eslintrc: `packages/${projectName}/eslint.config.js`,
+          patterns: ['.'],
+        })
+      : await eslintPlugin(await eslintConfigFromAllNxProjects()),
   ],
   categories: eslintCategories,
 });