+++ packages/core/src/lib/implementation/execute-plugin.ts
@@ -14,6 +14,7 @@ import {
   groupByStatus,
   logMultipleResults,
   pluralizeToken,
+  scoreAuditsWithTarget,
 } from '@code-pushup/utils';
 import {
   executePluginRunner,
@@ -57,6 +58,7 @@ export async function executePlugin(
     description,
     docsUrl,
     groups,
+    scoreTargets,
     ...pluginMeta
   } = pluginConfig;
   const { write: cacheWrite = false, read: cacheRead = false } = cache;
@@ -66,8 +68,8 @@ export async function executePlugin(
     ? // IF not null, take the result from cache
       ((await readRunnerResults(pluginMeta.slug, outputDir)) ??
       // ELSE execute the plugin runner
-      (await executePluginRunner(pluginConfig)))
-    : await executePluginRunner(pluginConfig);
+      (await executePluginRunner(pluginConfig, persist)))
+    : await executePluginRunner(pluginConfig, persist);
 
   if (cacheWrite) {
     await writeRunnerResults(pluginMeta.slug, outputDir, {
@@ -76,8 +78,13 @@ export async function executePlugin(
     });
   }
 
+  // transform audit scores to 1 when they meet/exceed their targets
+  const scoredAuditsWithTarget = scoreTargets
+    ? scoreAuditsWithTarget(audits, scoreTargets)
+    : audits;
+
   // enrich `AuditOutputs` to `AuditReport`
-  const auditReports: AuditReport[] = audits.map(
+  const auditReports: AuditReport[] = scoredAuditsWithTarget.map(
     (auditOutput: AuditOutput) => ({
       ...auditOutput,
       ...(pluginConfigAudits.find(