+++ testing/test-vitest-setup/src/configs/int.ts
@@ -1,23 +1,22 @@
 /// <reference types="vitest" />
-import { defineConfig, mergeConfig } from 'vite';
+import { defineConfig } from 'vitest/config';
 import { baseTestConfig } from './base.js';
 
-export const intTestConfig = mergeConfig(
-  baseTestConfig,
-  defineConfig({
-    test: {
-      coverage: {
-        reporter: ['text', 'lcov'],
-        reportsDirectory: './coverage',
-        exclude: ['mocks/**', '**/types.ts'],
-      },
-      include: ['src/**/*.int.test.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
-      setupFiles: [
-        '../../testing/test-setup/src/lib/console.mock.ts',
-        '../../testing/test-setup/src/lib/reset.mocks.ts',
-        '../../testing/test-setup/src/lib/portal-client.mock.ts',
-        '../../testing/test-setup/src/lib/extend/path.matcher.ts',
-      ],
+export const intTestConfig = defineConfig({
+  ...baseTestConfig,
+  test: {
+    ...baseTestConfig.test,
+    coverage: {
+      reporter: ['text', 'lcov'],
+      reportsDirectory: './coverage',
+      exclude: ['mocks/**', '**/types.ts'],
     },
-  }),
-);
+    include: ['src/**/*.int.test.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
+    setupFiles: [
+      '../../testing/test-setup/src/lib/console.mock.ts',
+      '../../testing/test-setup/src/lib/reset.mocks.ts',
+      '../../testing/test-setup/src/lib/portal-client.mock.ts',
+      '../../testing/test-setup/src/lib/extend/path.matcher.ts',
+    ],
+  },
+});