+++ testing/test-vitest-setup/src/configs/unit.ts
@@ -1,28 +1,27 @@
 /// <reference types="vitest" />
-import { defineConfig, mergeConfig } from 'vite';
+import { defineConfig } from 'vitest/config';
 import { baseTestConfig } from './base.js';
 
-export const unitTestConfig = mergeConfig(
-  baseTestConfig,
-  defineConfig({
-    test: {
-      coverage: {
-        reporter: ['text', 'lcov'],
-        reportsDirectory: './coverage',
-        exclude: ['mocks/**', '**/types.ts'],
-      },
-      include: ['src/**/*.unit.test.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
-      setupFiles: [
-        '../../testing/test-setup/src/lib/fs-with-cwd.setup-file.ts',
-        '../../testing/test-setup/src/lib/cliui.mock.ts',
-        '../../testing/test-setup/src/lib/git.mock.ts',
-        '../../testing/test-setup/src/lib/console.mock.ts',
-        '../../testing/test-setup/src/lib/reset.mocks.ts',
-        '../../testing/test-setup/src/lib/portal-client.mock.ts',
-        '../../testing/test-setup/src/lib/extend/ui-logger.matcher.ts',
-        '../../testing/test-setup/src/lib/extend/markdown-table.matcher.ts',
-        '../../testing/test-setup/src/lib/extend/jest-extended.matcher.ts',
-      ],
+export const unitTestConfig = defineConfig({
+  ...baseTestConfig,
+  test: {
+    ...baseTestConfig.test,
+    coverage: {
+      reporter: ['text', 'lcov'],
+      reportsDirectory: './coverage',
+      exclude: ['mocks/**', '**/types.ts'],
     },
-  }),
-);
+    include: ['src/**/*.unit.test.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
+    setupFiles: [
+      '../../testing/test-setup/src/lib/fs-with-cwd.setup-file.ts',
+      '../../testing/test-setup/src/lib/cliui.mock.ts',
+      '../../testing/test-setup/src/lib/git.mock.ts',
+      '../../testing/test-setup/src/lib/console.mock.ts',
+      '../../testing/test-setup/src/lib/reset.mocks.ts',
+      '../../testing/test-setup/src/lib/portal-client.mock.ts',
+      '../../testing/test-setup/src/lib/extend/ui-logger.matcher.ts',
+      '../../testing/test-setup/src/lib/extend/markdown-table.matcher.ts',
+      '../../testing/test-setup/src/lib/extend/jest-extended.matcher.ts',
+    ],
+  },
+});